backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
SQLAlchemy
pydantic
python-jose
passlib[bcrypt]
python-dotenv
